Break up some dependencies in the expression directory
