+++ libs/language-server/src/lib/ast/expressions/evaluation-strategy.ts
@@ -1,0 +1,8 @@
+// SPDX-FileCopyrightText: 2023 Friedrich-Alexander-Universitat Erlangen-Nurnberg
+//
+// SPDX-License-Identifier: AGPL-3.0-only
+
+export enum EvaluationStrategy {
+  EXHAUSTIVE,
+  LAZY,
+}
