+++ libs/language-server/src/lib/ast/expressions/operator-types.ts
@@ -1,0 +1,9 @@
+import {
+  BinaryExpression,
+  TernaryExpression,
+  UnaryExpression,
+} from '../generated/ast';
+
+export type UnaryExpressionOperator = UnaryExpression['operator'];
+export type BinaryExpressionOperator = BinaryExpression['operator'];
+export type TernaryExpressionOperator = TernaryExpression['operator'];
